🐞 Fix "dial tcp 10.245.0.3:25555: connect: no route to host"
The never-before-run nginx-release pipelines has unsurprisingly failed:
it mistakenly calls `start-bosh` in a sub-shell which has the
unfortunate consequence of tearing down the Director when the sub-shell
exits due to a cleverly-placed `trap EXIT`.

We now source the `start-bosh` command in the current shell, which is
the way it's intended to be used.

fixes <https://ci.bosh-ecosystem.cf-app.com/teams/main/pipelines/nginx-release/jobs/build/builds/1>:
```
Fetching info:
  Performing request GET 'https://10.245.0.3:25555/info':
    Performing GET request:
      Retry: Get "https://10.245.0.3:25555/info": dial tcp 10.245.0.3:25555: connect: no route to host
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>