Modules: headers-more: 0.30→0.33, upload 2.2.0→2.3.0
We bump two modules (`headers-more` and `nginx-upload-module`). We hope
that by bumping the upload module we no longer require the module patch,
which we have removed.

We also introduce a new module, `hmac_secure_link_module`, which is used
by the BOSH Director. We are attempting to converge the BOSH Director's
nginx release with the BOSH packages' nginx release.